+++ src/npmlicensecheck.ts
@@ -4,6 +4,13 @@ import * as exec from "@actions/exec";
 const blockedLicenses = core.getMultilineInput("blockedLicenses");
 const continueOnBlockedFound = core.getBooleanInput("continueOnBlockedFound");
 
+interface PackageEntry {
+  name: string;
+  license: string;
+  repository: string;
+  version: string;
+}
+
 const processNpm = async (projectPath: string): Promise<string> => {
   core.info(`Starting processNpm for: ${projectPath}`);
 
@@ -11,6 +18,34 @@ const processNpm = async (projectPath: string): Promise<string> => {
     silent: true,
   });
 
+  // create detailed report
+  const { stdout: licenseReportDetailed } = await exec.getExecOutput(
+    "npx",
+    ["license-compliance@2", "--production", "--format", "json", "--report", "detailed"],
+    { silent: true },
+  );
+
+  const tableData = (JSON.parse(licenseReportDetailed) as PackageEntry[]).map(({ name, version, license, repository }) => {
+    const status = blockedLicenses.includes(license) ? ":warning:" : ":white_check_mark:";
+    return [status, name, version, license, repository];
+  });
+
+  await core.summary
+    .addHeading("NPM license Details for " + projectPath)
+    // .addCodeBlock(licenseReportDetailed, "text")
+    .addTable([
+      [
+        { data: "Status", header: true },
+        { data: "Name", header: true },
+        { data: "Version", header: true },
+        { data: "License", header: true },
+        { data: "Repository", header: true },
+      ],
+      ...tableData,
+    ])
+    .write();
+
+  // create summary report for PR comment
   const { stdout: licenseReport } = await exec.getExecOutput(
     "npx",
     ["license-compliance@2", "--production", "--format", "json", "--report", "summary"],