+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added npm license report to pipeline
+
+### Changed
+
+- npm packages search includes root folder
+
 ## [0.3.0] - 2023-11-02
 
 ### Added